Refactor DraftPrescription list
* show some coverage info
