+++ packages/components/src/systems/DraftPrescriptions/index.tsx
@@ -1,3 +1,3 @@
-export * from './DraftPrescriptions';
+export * from './DraftPrescriptionList';
 export * from './DraftPrescriptionsProvider';
-export type { DraftPrescription } from './DraftPrescriptions';
+export type { DraftPrescription } from './DraftPrescriptionList';
